refactoring and cleaning the protocols
refactoring and cleaning the protocols
- putting HTML report into DesktopProtocol and overwriting preSelectAction
- cleaning the other protocols
